Remove code duplication in both FindingsReports
